requirements.txt
streamlit
openai
reportlab
python-dotenv
folium
streamlit-folium
